HTML
Added some styling, so that the images sit in a 9x9 column. Each img is
assigned a “position” which represents its rightful position. Each div
where an img is placed in has a “position” number as well. The image is
correctly set when each image position matches the parent div’s
position.